Add API, refresh DB-layout (#12)
Re-think database model and start creating API.
old databases are NOT compatible

* Add more data for database population

* Re-worked usermodel

* Add UUID to user (inter-application ID)

To be used for example in identifying the user in BBB, which makes
managing our meetings easier.

* Fix admin register

* Improving roulette models, including new matching

The new matching now consists of two seperate requests which can be
matched. Once matched a Match of those will be created, completing the
user_affirm. This means that during a running request, multiple Matches
(in sequence) are possible.
Right now the model is limited to one Match per Request at a time,
should we consider adding multiple?

* Work on making custom user model work in django

* Fix populate_database

* Fix models (constraints can't contain joins)

* Fix application name in configs

* Fix up models for Admin and in general

* Make unalterable

* add landing module for e-mail collection

* add missing migrations

* add landing/add_individual

* Change studentdata, tutordata reference to user

* add landing/add_individual

* Work on account-api

* Add knox-auth to requirements

* Add SchemaView dependencies

* Revamp connection StudentData,TutorData to CustomUser

* Work on API

* Add API documentation view /swagger-ui

* add swagger-ui

* Add account-create, cleanup schemas

* clean up urls and landing-model